renames the table to PaginatedTable
